+++ hypervisor/common/hypercall.c
@@ -924,72 +924,6 @@ hcall_reset_ptdev_intr_info(struct acrn_vm *vm, uint16_t vmid, uint64_t param)
 	return ret;
 }
 
-#ifdef HV_DEBUG
-/**
- * @brief Setup a share buffer for a VM.
- *
- * @param vm Pointer to VM data structure
- * @param param guest physical address. This gpa points to
- *              struct sbuf_setup_param
- *
- * @pre Pointer vm shall point to VM0
- * @return 0 on success, non-zero on error.
- */
-int32_t hcall_setup_sbuf(struct acrn_vm *vm, uint64_t param)
-{
-	struct sbuf_setup_param ssp;
-	uint64_t *hva;
-
-	(void)memset((void *)&ssp, 0U, sizeof(ssp));
-
-	if (copy_from_gpa(vm, &ssp, param, sizeof(ssp)) != 0) {
-		pr_err("%s: Unable copy param to vm\n", __func__);
-		return -1;
-	}
-
-	if (ssp.gpa != 0U) {
-		hva = (uint64_t *)gpa2hva(vm, ssp.gpa);
-	} else {
-		hva = (uint64_t *)NULL;
-	}
-
-	return sbuf_share_setup(ssp.pcpu_id, ssp.sbuf_id, hva);
-}
-#endif
-
-#ifdef HV_DEBUG
-/**
-  * @brief Setup the hypervisor NPK log.
-  *
-  * @param vm Pointer to VM data structure
-  * @param param guest physical address. This gpa points to
-  *              struct hv_npk_log_param
-  *
-  * @pre Pointer vm shall point to VM0
-  * @return 0 on success, non-zero on error.
-  */
-int32_t hcall_setup_hv_npk_log(struct acrn_vm *vm, uint64_t param)
-{
-	struct hv_npk_log_param npk_param;
-
-	(void)memset((void *)&npk_param, 0U, sizeof(npk_param));
-
-	if (copy_from_gpa(vm, &npk_param, param, sizeof(npk_param)) != 0) {
-		pr_err("%s: Unable copy param from vm\n", __func__);
-		return -1;
-	}
-
-	npk_log_setup(&npk_param);
-
-	if (copy_to_gpa(vm, &npk_param, param, sizeof(npk_param)) != 0) {
-		pr_err("%s: Unable copy param to vm\n", __func__);
-		return -1;
-	}
-
-	return 0;
-}
-#endif
-
 /**
  * @brief Get VCPU Power state.
  *
