+++ hypervisor/include/common/hypercall.h
@@ -307,44 +307,18 @@ int32_t hcall_set_ptdev_intr_info(struct acrn_vm *vm, uint16_t vmid, uint64_t pa
 int32_t hcall_reset_ptdev_intr_info(struct acrn_vm *vm, uint16_t vmid,
 	uint64_t param);
 
-#ifdef HV_DEBUG
 /**
- * @brief Setup a share buffer for a VM.
- *
- * @param vm Pointer to VM data structure
- * @param param guest physical address. This gpa points to
- *              struct sbuf_setup_param
- *
- * @pre Pointer vm shall point to VM0
- * @return 0 on success, non-zero on error.
- */
-int32_t hcall_setup_sbuf(struct acrn_vm *vm, uint64_t param);
-
-/**
-  * @brief Setup the hypervisor NPK log.
+  * @brief Setup hypervisor debug infrastructure, such as share buffer, NPK log and profiling.
   *
   * @param vm Pointer to VM data structure
-  * @param param guest physical address. This gpa points to
-  *              struct hv_npk_log_param
+  * @param param1 hypercall param1 from guest
+  * @param param2 hypercall param2 from guest
+  * @param hypcall_id hypercall ID from guest
   *
   * @pre Pointer vm shall point to VM0
   * @return 0 on success, non-zero on error.
   */
-int32_t hcall_setup_hv_npk_log(struct acrn_vm *vm, uint64_t param);
-#endif
-
-/**
- * @brief Execute profiling operation
- *
- * @param vm Pointer to VM data structure
- * @param cmd profiling command to be executed
- * @param cmd profiling command to be executed
- * @param param guest physical address. This gpa points to
- *             data structure required by each command
- *
- * @return 0 on success, non-zero on error.
- */
-int32_t hcall_profiling_ops(struct acrn_vm *vm, uint64_t cmd, uint64_t param);
+int32_t hcall_debug(struct acrn_vm *vm, uint64_t param1, uint64_t param2, uint64_t hypcall_id);
 
 /**
  * @brief Get VCPU Power state.
